% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ip2location.lite.db1}
\alias{ip2location.lite.db1}
\title{IP database - small}
\format{A data.frame with four variables
\describe{
  \item{IPfrom}{Start of IP integer range}
  \item{IPto}{End of IP integer range}
  \item{Abrv}{Two-character country code based on ISO 3166.}
  \item{Country}{Country name based on ISO 3166.}
}}
\source{
\url{http://lite.ip2location.com}
}
\usage{
ip2location.lite.db1
}
\description{
A dataset containing 149,449 IP integer ranges and their corresponding country
names and abbreviations
}
\details{
Last updated: 06/06/2016
}
\keyword{datasets}

