% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{ip2location.lite.db11}
\alias{ip2location.lite.db11}
\title{IP database - large}
\format{A data.frame with ten variables
\describe{
  \item{IPfrom}{Start of IP integer range}
  \item{IPto}{End of IP integer range}
  \item{Abrv}{Two-character country code based on ISO 3166.}
  \item{Country}{Country name based on ISO 3166.}
  \item{Region}{Region within country}
  \item{City}{City within region}
  \item{Lat}{Latitude}
  \item{Long}{Longitude}
  \item{Zip}{Zip code}
  \item{GMT}{Greenwich Mean Time}
}}
\source{
\url{http://lite.ip2location.com}
}
\usage{
ip2location.lite.db11
}
\description{
A dataset containing 3,832,909 IP integer ranges and their corresponding country
names, abbreviations, regions, cities, latitudes, longitudes, zip codes, and Greenwich Mean Times
}
\details{
Last updated: 06/02/2016
}
\keyword{datasets}

