% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_binary.R
\name{IP_binary}
\alias{IP_binary}
\title{Convert IP address to binary}
\usage{
IP_binary(IP.address, as.list = FALSE)
}
\arguments{
\item{IP.address}{A character vector of IP addresses}

\item{as.list}{logical, \code{FALSE} returns a named character vector instead of a list}
}
\value{
A list of IP addresses converted to character strings of binary
  values
}
\description{
Create method for splitting an IP address values into lists of binary values
}

