% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_country.R
\name{IP_country}
\alias{IP_country}
\title{Convert IP address to country name}
\usage{
IP_country(IP.address, IP.database = NULL)
}
\arguments{
\item{IP.address}{a character or factor vector of one or more IP addresses}

\item{IP.database}{an IP database, see \code{?ip2location.lite.db1}}
}
\value{
Returns a factor vector of country names corresponding to
  \code{IP.address}
}
\description{
Convert IP address to country name
}
\examples{
IP_country(IPs)

}

