% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_generator.R
\name{IP_generator}
\alias{IP_generator}
\title{Random IP address generator}
\usage{
IP_generator(n)
}
\arguments{
\item{n}{number of IP addresses to return}
}
\value{
Returns a random IP address vector of length \code{n}
}
\description{
Random IP address generator
}
\examples{
# Generate 100 random IP addresses
IP_generator(100)
}

