% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_country.R
\name{IP_integer}
\alias{IP_integer}
\title{Convert IP address to IP integer}
\usage{
IP_integer(IP.address)
}
\arguments{
\item{IP.address}{a character or factor vector of one or more IP addresses}
}
\value{
Returns a numeric vector of IP integers from \code{IP.address}. The
  coversion formula used here splits the string into four octets, multiplies
  the first three by 256^(4-n), and takes the sum of the four modified
  octets.
}
\description{
Convert IP address to IP integer
}
\examples{
head(IP_integer(IPs))

}

