% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_country.R
\name{IP_location}
\alias{IP_location}
\title{Convert IP address to location (country name, region, city, zip code,
latitude, longitude, and GMT)}
\usage{
IP_location(IP.address, IP.database = NULL)
}
\arguments{
\item{IP.address}{a character or factor vector of one or more IP addresses}

\item{IP.database}{an IP database, see \code{?ip2location.lite.db11}}
}
\value{
Returns a data.frame containin the country name, region, city, and
  zip code corresponding to \code{IP.address}
}
\description{
Convert IP address to location (country name, region, city, zip code,
latitude, longitude, and GMT)
}
\examples{
IP_location(IPs)

}

