% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_country.R
\name{IP_lookup}
\alias{IP_lookup}
\title{Match IP integer to country name}
\usage{
IP_lookup(IP.integer, IP.database = NULL)
}
\arguments{
\item{IP.integer}{a character or factor vector of one or more IP addresses}

\item{IP.database}{an IP database, see \code{?ip2location.lite.db1} for details on
default database from \url{http://lite.ip2location.com}}
}
\value{
Returns country from IP integers
}
\description{
Match IP integer to country name
}
\examples{
IP.integer <- IP_integer(IPs)
head(IP_lookup(IP.integer))
}

