% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_country.R
\name{IP_lookup_full}
\alias{IP_lookup_full}
\title{Match IP integer to all location data}
\usage{
IP_lookup_full(IP.integer, IP.database = NULL)
}
\arguments{
\item{IP.integer}{a character or factor vector of one or more IP addresses}

\item{IP.database}{an IP database, see \code{?ip2location} for details on
default database from \url{http://lite.ip2location.com}}
}
\value{
Returns country from IP integers
}
\description{
Match IP integer to all location data
}
\examples{
IP.integer <- IP_integer(IPs)
IP_lookup_full(IP.integer)
}

