% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_plot.R
\name{IP_plot}
\alias{IP_plot}
\title{Convert IPs to countries and plot on world map}
\usage{
IP_plot(IP.address)
}
\arguments{
\item{IP.address}{a vector or column of IP addresses}
}
\value{
Returns a world map plot with gradient coloring reflecting the
  percentage of IP addresses originating in each country
}
\description{
Convert IPs to countries and plot on world map
}
\examples{
IP_plot(IPs)
}

