% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_country.R
\name{IP_split}
\alias{IP_split}
\title{Split IP addresses}
\usage{
IP_split(IP.address, integer = TRUE, data.frame = TRUE)
}
\arguments{
\item{IP.address}{vector of IPv4 addresses}

\item{integer}{logical, convert output to class integer}

\item{data.frame}{logical, convert output to data.frame}
}
\value{
Returns a matrix or data.frame with four columns and as many rows as
  IP.addresses were inputted
}
\description{
Split IP addresses
}
\examples{
head(IP_split(IPs))
}

