% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{IPs}
\alias{IPs}
\title{IP addresses}
\format{A character vector with length 1000}
\source{
Generated using \code{\link{IP_generator}}
}
\usage{
IPs
}
\description{
A dataset containing 1000 randomly generated IP addresses
}
\keyword{datasets}

